requirements.txt
Flask
requests
Pillow
werkzeug
mysql-connector-python
python-dotenv